# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Blog:
    properties:
      author:
        type: string
      content:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  models.Login:
    properties:
      password:
        type: string
      userName:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: mail@necmettincimen.com.tr
    name: API Support
    url: https://necmettincimen.com.tr
  description: This is a sample go server blog api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog Api Golang
  version: "1.0"
paths:
  /blog:
    delete:
      consumes:
      - application/json
      parameters:
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.Blog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Blog'
      tags:
      - blog
    get:
      consumes:
      - application/json
      parameters:
      - description: model
        in: query
        name: take
        required: true
        type: integer
      - description: model
        in: query
        name: skip
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Blog'
            type: array
      tags:
      - blog
    post:
      consumes:
      - application/json
      parameters:
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.Blog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Blog'
      tags:
      - blog
    put:
      consumes:
      - application/json
      parameters:
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.Blog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Blog'
      tags:
      - blog
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Login'
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"